# Optimize Scala Future -> Twitter Future conversion