--- a/core/src/main/scala/io/finch/syntax/scala/CallingThreadExecutionContext.scala
+++ b/core/src/main/scala/io/finch/syntax/scala/CallingThreadExecutionContext.scala
@@ -1,0 +1,9 @@
+package io.finch.syntax.scala
+
+import scala.concurrent.ExecutionContext
+
+private[scala] object CallingThreadExecutionContext extends ExecutionContext {
+  def execute(runnable: Runnable): Unit = runnable.run()
+
+  def reportFailure(throwable: Throwable): Unit = throwable.printStackTrace()
+}