@ImLiar 2017-11-04
But now user can not provide his own EC, right?
May be it would be better to mention this trick in documentation and leave it for users to decide.