@shado23 2017-11-04
> But now user can not provide his own EC, right?
May be it would be better to mention this trick in documentation and leave it for users to decide.

The CallingThreadExecutionContext is only used for the conversion step, the Future itself will still be scheduled on whatever ExecutionContext the user provides. (In fact it would be a bad idea to force the future to run on the calling thread as it could be long running/blocking.) Essentially the conversion step will be run on whatever thread the future finishes executing on.

As far as I know there's no point in offering a choice of ExecutionContext for the conversion step, as any other choice is slower and there's no possible benefit of forcing an asynchronous boundary here. The conversion step itself is very simple (which is why the scheduling overhead of the future cause such a large performance hit).

This optimization is used for similar purposes by [futiles](https://github.com/johanandren/futiles/blob/master/src/main/scala/markatta/futiles/CallingThreadExecutionContext.scala) and [swave](https://github.com/sirthias/swave/blob/master/core/src/main/scala/swave/core/impl/CallingThreadExecutionContext.scala) and is well established.