@shado23 2017-11-06
I've added another commit special casing already completed futures. This yields an additional slight performance increase:
```
Running 30s test @ http://127.0.0.1:8080/health
  6 threads and 6 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   142.82us  470.46us  14.59ms   97.88%
    Req/Sec    11.22k     2.89k   17.18k    69.29%
  2010878 requests in 30.10s, 172.60MB read
Requests/sec:  66807.86
Transfer/sec:      5.73MB
```