@shado23 2017-11-11
> Yeah, @vkostyukov this looks good to me. Its a nice bump in performance. Thanks @shado23!

Your're welcome.